Task5

+++ Tyuiu.MazurkevichVS.Sprint0.Task4.V0.Lib/DataService.cs
@@ -14,7 +14,7 @@
         {
             return a * b;
         }
-        public static int Division(int a, int b)
+        public static double Division(int a, int b)
         {
             return a / b;
         }
